requirements.txt
apex==0.9.10dev
diffusers==0.30.3
h5py==2.8.0
h5py==3.10.0
jsonlines==4.0.0
line_profiler==4.1.3
matplotlib==3.2.2
networkx==2.8.4
numpy==1.15.1
numpy==1.22.3
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==9.0.1
Pillow==9.1.0
Pillow==10.4.0
progressbar33==2.4
scikit_learn==1.5.2
tensorboardX==2.6.2.2
timm==1.0.9
torch==1.11.0
tqdm==4.62.3
transformers==4.45.1
